pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["yijun"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
gunicorn = "^22.0.0"
sqlalchemy = "^2.0.29"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
flask = "^3.0.3"
pydantic = "^2.7.1"
myauth = {git = "git@github.com:yijunx/auth", rev = "0.0.1"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.2"
isort = "^5.13.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
